+++ src/background/main.rs
@@ -6,6 +6,7 @@ mod cli;
 mod error_handler;
 mod exposed;
 mod hook;
+mod monitor;
 mod plugins;
 mod seelen;
 mod seelen_bar;
@@ -13,13 +14,14 @@ mod seelen_shell;
 mod seelen_weg;
 mod seelen_wm;
 mod state;
+mod system;
 mod tray;
 mod utils;
 mod windows_api;
 mod winevent;
-mod system;
 
 use cli::handle_cli_info;
+use color_eyre::owo_colors::OwoColorize;
 use error_handler::Result;
 use exposed::register_invoke_handler;
 use plugins::register_plugins;
@@ -32,9 +34,32 @@ use crate::cli::SEELEN_COMMAND_LINE;
 fn main() -> Result<()> {
     color_eyre::install().expect("Failed to install color_eyre");
     std::panic::set_hook(Box::new(|info| {
-        if let Some(s) = info.payload().downcast_ref::<String>() {
-            log::error!("{}", s);
+        let cause = info
+            .payload()
+            .downcast_ref::<String>()
+            .map(|s| s.to_string())
+            .unwrap_or_else(|| {
+                info.payload()
+                    .downcast_ref::<&str>()
+                    .unwrap_or(&"<cause unknown>")
+                    .to_string()
+            });
+
+        let mut string_location = String::from("<location unknown>");
+        if let Some(location) = info.location() {
+            string_location = format!(
+                "{}:{}:{}",
+                location.file(),
+                location.line(),
+                location.column()
+            );
         }
+
+        log::error!(
+            "A panic occurred:\n  Cause: {}\n  Location: {}",
+            cause.cyan(),
+            string_location.purple()
+        );
     }));
 
     let command = SEELEN_COMMAND_LINE.lock().clone();
@@ -51,7 +76,7 @@ fn main() -> Result<()> {
     let app = app_builder
         .setup(move |app| {
             log::info!("───────────────────── Starting Seelen ─────────────────────");
-            let mut seelen = unsafe { SEELEN.make_guard_unchecked()};
+            let mut seelen = unsafe { SEELEN.make_guard_unchecked() };
             seelen.init(app.handle().clone())?;
 
             handle_tray_icon(app)?;